3lab

+++ LibraryCPP/list.h
@@ -3,7 +3,7 @@
 
 // List
 // Stores integer values inside
-typedef double Data;
+typedef int Data;
 
 struct List;
 struct ListItem;
